+++ CHANGELOG.md
@@ -6,8 +6,9 @@
     * **Note** Add retry logic for retryable HTTP status codes. [#147](https://github.com/patrickcping/pingone-go-sdk-v2/pull/147)
 * `github.com/patrickcping/pingone-go-sdk-v2/authorize` : [v0.1.5](./authorize/CHANGELOG.md)
     * **Note** Add retry logic for retryable HTTP status codes. [#147](https://github.com/patrickcping/pingone-go-sdk-v2/pull/147)
-* `github.com/patrickcping/pingone-go-sdk-v2/management` : [v0.18.1](./management/CHANGELOG.md)
+* `github.com/patrickcping/pingone-go-sdk-v2/management` : [v0.19.0](./management/CHANGELOG.md)
     * **Note** Add retry logic for retryable HTTP status codes. [#147](https://github.com/patrickcping/pingone-go-sdk-v2/pull/147)
+    * **Enhancement** Added support for the Google Play Integrity API, `GooglePlay` for the `ApplicationOIDC` object model.  *Breaking change note*: This is now a required attribute when configuring integrity detection on Android devices.
 * `github.com/patrickcping/pingone-go-sdk-v2/mfa` : [v0.9.2](./mfa/CHANGELOG.md)
     * **Note** Add retry logic for retryable HTTP status codes. [#147](https://github.com/patrickcping/pingone-go-sdk-v2/pull/147)
 * `github.com/patrickcping/pingone-go-sdk-v2/risk` : [v0.5.0](./risk/CHANGELOG.md)